+++ src/js/ConfigMenu.js
@@ -1,7 +1,7 @@
 import $ from "jquery";
 import MarkdownIt from "markdown-it";
 import MarkdownItAttrs from "markdown-it-attrs";
-import markdownItBracketedSpans from "markdown-it-bracketed-spans";
+import MarkdownItBracketedSpans from "markdown-it-bracketed-spans";
 
 import svgUndo from "svg/undo";
 
@@ -82,8 +82,8 @@ export default class ConfigMenu {
             linkify: true,
             typographer: true
         });
+        this.#md.use(MarkdownItBracketedSpans);
         this.#md.use(MarkdownItAttrs);
-        this.#md.use(markdownItBracketedSpans);
 
         const container = document.createElement("div");
         container.id = "dribbblish-config";